Tweak the type docs.
Refer to them as "type constraints" as discussed on the syntax list to differenciate from
Moose types.

Give a clear layout of the logic for resolving a type.

Don't give away the order of type resolution, just leave it as undeclared so nobody gets
any ideas (or if they do and it changes it's their problem).