5.14.1 sometimes doesn't know it's dead.
A strict error will not cause a compilation error, but it will be in $@ and will cause any
further uses of eval STRING (including require) to fail.  This means strict errors can show
up as failures to load PPI or other unrelated things.  This code checks to see if the parser
is in a bad state before proceeding, and erroring out with whatever is in $@ if it's fucked.