Demonstrate that we can compile methods at compile time without BeginLift.
Rather than using Devel::BeginLift or declaring things with a BEGIN block,
we simply rewrite "func foo {...}" to be "sub foo {...}" and the rest
happens magically.

The one caveat is Devel::Declare expects `func` and `method` to get
the code ref being generated.  Rather than fight this, we go along
with it.

Only four tests fail.  Fixes on the way.