first cut at adding a //= default
It adds a second kind of default to Method::Signatures,
so that instead of:

    func foo ($opt_arg?) {
        $opt_arg //= $DEFAULT;
        ...
    }

you can just write:

    func foo ($opt_arg //= $DEFAULT) {
        ...
    }

I submit it for your consideration, with the full disclosure that we
have not approved such a mechanism for Perl 6. And that Larry may well
never allow it. Personally, I find it natural and obvious, but I leave the
decision entirely up to you both. >;-)