changing //= default to "when undef"
The "when * < 0" syntax is borrowed from Perl 6, where it's a standard and
widely used feature. But I agree that it might be a little too funky for
MS. See the attached next-generation patch, which uses a Perl 5
compatible block-and-$_ syntax instead. Much more appropriate, I think.

Note that, compared to the previous submission, the attached patch also
has more robust parsing of 'when' modifiers.