added the yada yada pseudo-parameter
For example, please find attached another patch that implements a
proposed solution to the issue under discussion here:

    https://github.com/schwern/method-signatures/issues/49

For once, it's not stolen from Perl 6, but rather from Vimscript.
It implements a '...' pseudo-parameter (known as "yada yada")
that simply turns off all checking for extraneous arguments.

I think it's a clean, readable, and "intentional" solution to the
problem and would encourage both of you to consider it.