Add the line number to the Parameter object.
Because the PPI tokens have the line number informaiton, this required
unifying the parameter splitting with building the Parameter object.
This is a good idea, the intermediate step was unnecessary, but I'm
not happy with how complicated the loop has gotten.

Ideally, MS::Parameter->new would be passed the PPI tokens and be left
to do the work that _build_parameters is currently doing.  This would
also avoid having to reconstruct the code and then reparse it with PPI
inside Parameter.  The problem is how to neatly snip off a fragment
of a PPI::Statement.  Not ready to tackle that.

Haven't tested the line number yet.  That's coming next.