Eliminate a now unnecessary local $@.
This was causing failure on 5.12, 5.10 and 5.8.  I don't know why it's version
specific.  It was introduced in e0cd9bc9 but the code it's in has been shifted
around so it's now happening during object creation.  The tests introduced at
that time still work.

It's possible this was a bad interaction with PPI which uses eval without localizing
$@.