Handle the case where the signature contains a valid $,
The only time "$," must be broken up into "$" and "," is when it's the first
thing in a signature's parameter, which can also happen if it's only seen
whitespace up to the current param's parse.  If there are any non-whitespace
tokens already seen in the current param we're working on, then this "$,"
must be a real "$,".

The only other case to consider is something like
$foo , $, , $bar
which would mean the second param is to be called "$,".  It's not possible
since you can't have a lexical var called $,