Allow function attributes to come before the signature/prototype
Override strip_attrs() from Devel::Declare::Context::Simple to fix a problem
where it'll get stuck in an endless loop looking for the open brace "{".
Stops when it finds an open paren "(" as well.

Override parser() from D::D::MethodInstaller::Simple to call strip_proto()
and strip_attrs() in the same order the signature/prototype and attributes
appear in the source.